Move group_experiments to the _group file in the experiment module.
