+++ pyquil/tests/test_operator_estimation.py
@@ -20,59 +20,19 @@ from pyquil.experiment import (
     plusX,
     plusY,
     plusZ,
-    zeros_state,
 )
 from pyquil.gates import CNOT, CZ, H, I, RX, RY, RZ, X, Y
 from pyquil.operator_estimation import (
     _one_q_sic_prep,
-    _max_weight_operator,
-    _max_weight_state,
-    _max_tpb_overlap,
     group_experiments,
     measure_observables,
     _ops_bool_to_prog,
     _stats_from_measurements,
     ratio_variance,
     _calibration_program,
-    _pauli_to_product_state,
 )
-from pyquil.paulis import sI, sX, sY, sZ, PauliSum, PauliTerm
+from pyquil.paulis import sI, sX, sY, sZ, PauliSum
 from pyquil.quilbase import Pragma
-
-
-@pytest.fixture(params=["clique-removal", "greedy"])
-def grouping_method(request):
-    return request.param
-
-
-def test_expt_settings_share_ntpb():
-    expts = [
-        [
-            ExperimentSetting(zeros_state([0, 1]), sX(0) * sI(1)),
-            ExperimentSetting(zeros_state([0, 1]), sI(0) * sX(1)),
-        ],
-        [
-            ExperimentSetting(zeros_state([0, 1]), sZ(0) * sI(1)),
-            ExperimentSetting(zeros_state([0, 1]), sI(0) * sZ(1)),
-        ],
-    ]
-    for group in expts:
-        for e1, e2 in itertools.combinations(group, 2):
-            assert _max_weight_state([e1.in_state, e2.in_state]) is not None
-            assert _max_weight_operator([e1.out_operator, e2.out_operator]) is not None
-
-
-def test_group_experiments(grouping_method):
-    expts = [  # cf above, I removed the inner nesting. Still grouped visually
-        ExperimentSetting(TensorProductState(), sX(0) * sI(1)),
-        ExperimentSetting(TensorProductState(), sI(0) * sX(1)),
-        ExperimentSetting(TensorProductState(), sZ(0) * sI(1)),
-        ExperimentSetting(TensorProductState(), sI(0) * sZ(1)),
-    ]
-    suite = TomographyExperiment(expts, Program())
-    grouped_suite = group_experiments(suite, method=grouping_method)
-    assert len(suite) == 4
-    assert len(grouped_suite) == 2
 
 
 def test_measure_observables(forest):
@@ -163,176 +123,6 @@ def test_no_complex_coeffs(forest):
     )
     with pytest.raises(ValueError):
         list(measure_observables(qc, suite, n_shots=2000))
-
-
-def test_max_weight_operator_1():
-    pauli_terms = [sZ(0), sX(1) * sZ(0), sY(2) * sX(1)]
-    assert _max_weight_operator(pauli_terms) == sY(2) * sX(1) * sZ(0)
-
-
-def test_max_weight_operator_2():
-    pauli_terms = [sZ(0), sX(1) * sZ(0), sY(2) * sX(1), sZ(5) * sI(3)]
-    assert _max_weight_operator(pauli_terms) == sZ(5) * sY(2) * sX(1) * sZ(0)
-
-
-def test_max_weight_operator_3():
-    pauli_terms = [sZ(0) * sX(5), sX(1) * sZ(0), sY(2) * sX(1), sZ(5) * sI(3)]
-    assert _max_weight_operator(pauli_terms) is None
-
-
-def test_max_weight_operator_misc():
-    assert _max_weight_operator([sZ(0), sZ(0) * sZ(1)]) is not None
-    assert _max_weight_operator([sX(5), sZ(4)]) is not None
-    assert _max_weight_operator([sX(0), sY(0) * sZ(2)]) is None
-
-    x_term = sX(0) * sX(1)
-    z1_term = sZ(1)
-    z0_term = sZ(0)
-    z0z1_term = sZ(0) * sZ(1)
-    assert _max_weight_operator([x_term, z1_term]) is None
-    assert _max_weight_operator([z0z1_term, x_term]) is None
-
-    assert _max_weight_operator([z1_term, z0_term]) is not None
-    assert _max_weight_operator([z0z1_term, z0_term]) is not None
-    assert _max_weight_operator([z0z1_term, z1_term]) is not None
-    assert _max_weight_operator([z0z1_term, sI(1)]) is not None
-    assert _max_weight_operator([z0z1_term, sI(2)]) is not None
-    assert _max_weight_operator([z0z1_term, sX(5) * sZ(7)]) is not None
-
-    xxxx_terms = (
-        sX(1) * sX(2)
-        + sX(2)
-        + sX(3) * sX(4)
-        + sX(4)
-        + sX(1) * sX(3) * sX(4)
-        + sX(1) * sX(4)
-        + sX(1) * sX(2) * sX(3)
-    )
-    true_term = sX(1) * sX(2) * sX(3) * sX(4)
-    assert _max_weight_operator(xxxx_terms.terms) == true_term
-
-    zzzz_terms = sZ(1) * sZ(2) + sZ(3) * sZ(4) + sZ(1) * sZ(3) + sZ(1) * sZ(3) * sZ(4)
-    assert _max_weight_operator(zzzz_terms.terms) == sZ(1) * sZ(2) * sZ(3) * sZ(4)
-
-    pauli_terms = [sZ(0), sX(1) * sZ(0), sY(2) * sX(1), sZ(5) * sI(3)]
-    assert _max_weight_operator(pauli_terms) == sZ(5) * sY(2) * sX(1) * sZ(0)
-
-
-def test_max_weight_operator_4():
-    # this last example illustrates that a pair of commuting operators
-    # need not be diagonal in the same tpb
-    assert _max_weight_operator([sX(1) * sZ(0), sZ(1) * sX(0)]) is None
-
-
-def test_max_weight_state_1():
-    states = [plusX(0) * plusZ(1), plusX(0), plusZ(1)]
-    assert _max_weight_state(states) == states[0]
-
-
-def test_max_weight_state_2():
-    states = [plusX(1) * plusZ(0), plusX(0), plusZ(1)]
-    assert _max_weight_state(states) is None
-
-
-def test_max_weight_state_3():
-    states = [plusX(0) * minusZ(1), plusX(0), minusZ(1)]
-    assert _max_weight_state(states) == states[0]
-
-
-def test_max_weight_state_4():
-    states = [plusX(1) * minusZ(0), plusX(0), minusZ(1)]
-    assert _max_weight_state(states) is None
-
-
-def test_max_tpb_overlap_1():
-    tomo_expt_settings = [
-        ExperimentSetting(plusZ(1) * plusX(0), sY(2) * sY(1)),
-        ExperimentSetting(plusX(2) * plusZ(1), sY(2) * sZ(0)),
-    ]
-    tomo_expt_program = Program(H(0), H(1), H(2))
-    tomo_expt = TomographyExperiment(tomo_expt_settings, tomo_expt_program)
-    expected_dict = {
-        ExperimentSetting(plusX(0) * plusZ(1) * plusX(2), sZ(0) * sY(1) * sY(2)): [
-            ExperimentSetting(plusZ(1) * plusX(0), sY(2) * sY(1)),
-            ExperimentSetting(plusX(2) * plusZ(1), sY(2) * sZ(0)),
-        ]
-    }
-    assert expected_dict == _max_tpb_overlap(tomo_expt)
-
-
-def test_max_tpb_overlap_2():
-    expt_setting = ExperimentSetting(
-        _pauli_to_product_state(PauliTerm.from_compact_str("(1+0j)*Z7Y8Z1Y4Z2Y5Y0X6")),
-        PauliTerm.from_compact_str("(1+0j)*Z4X8Y5X3Y7Y1"),
-    )
-    p = Program(H(0), H(1), H(2))
-    tomo_expt = TomographyExperiment([expt_setting], p)
-    expected_dict = {expt_setting: [expt_setting]}
-    assert expected_dict == _max_tpb_overlap(tomo_expt)
-
-
-def test_max_tpb_overlap_3():
-    # add another ExperimentSetting to the above
-    expt_setting = ExperimentSetting(
-        _pauli_to_product_state(PauliTerm.from_compact_str("(1+0j)*Z7Y8Z1Y4Z2Y5Y0X6")),
-        PauliTerm.from_compact_str("(1+0j)*Z4X8Y5X3Y7Y1"),
-    )
-    expt_setting2 = ExperimentSetting(plusZ(7), sY(1))
-    p = Program(H(0), H(1), H(2))
-    tomo_expt2 = TomographyExperiment([expt_setting, expt_setting2], p)
-    expected_dict2 = {expt_setting: [expt_setting, expt_setting2]}
-    assert expected_dict2 == _max_tpb_overlap(tomo_expt2)
-
-
-def test_group_experiments_greedy():
-    ungrouped_tomo_expt = TomographyExperiment(
-        [
-            [
-                ExperimentSetting(
-                    _pauli_to_product_state(PauliTerm.from_compact_str("(1+0j)*Z7Y8Z1Y4Z2Y5Y0X6")),
-                    PauliTerm.from_compact_str("(1+0j)*Z4X8Y5X3Y7Y1"),
-                )
-            ],
-            [ExperimentSetting(plusZ(7), sY(1))],
-        ],
-        program=Program(H(0), H(1), H(2)),
-    )
-    grouped_tomo_expt = group_experiments(ungrouped_tomo_expt, method="greedy")
-    expected_grouped_tomo_expt = TomographyExperiment(
-        [
-            [
-                ExperimentSetting(
-                    TensorProductState.from_str(
-                        "Z0_7 * Y0_8 * Z0_1 * Y0_4 * Z0_2 * Y0_5 * Y0_0 * X0_6"
-                    ),
-                    PauliTerm.from_compact_str("(1+0j)*Z4X8Y5X3Y7Y1"),
-                ),
-                ExperimentSetting(plusZ(7), sY(1)),
-            ]
-        ],
-        program=Program(H(0), H(1), H(2)),
-    )
-    assert grouped_tomo_expt == expected_grouped_tomo_expt
-
-
-def test_expt_settings_diagonal_in_tpb():
-    def _expt_settings_diagonal_in_tpb(es1: ExperimentSetting, es2: ExperimentSetting):
-        """
-        Extends the concept of being diagonal in the same tpb to ExperimentSettings, by
-        determining if the pairs of in_states and out_operators are separately diagonal in the same
-        tpb
-        """
-        max_weight_in = _max_weight_state([es1.in_state, es2.in_state])
-        max_weight_out = _max_weight_operator([es1.out_operator, es2.out_operator])
-        return max_weight_in is not None and max_weight_out is not None
-
-    expt_setting1 = ExperimentSetting(plusZ(1) * plusX(0), sY(1) * sZ(0))
-    expt_setting2 = ExperimentSetting(plusY(2) * plusZ(1), sZ(2) * sY(1))
-    assert _expt_settings_diagonal_in_tpb(expt_setting1, expt_setting2)
-    expt_setting3 = ExperimentSetting(plusX(2) * plusZ(1), sZ(2) * sY(1))
-    expt_setting4 = ExperimentSetting(plusY(2) * plusZ(1), sX(2) * sY(1))
-    assert not _expt_settings_diagonal_in_tpb(expt_setting2, expt_setting3)
-    assert not _expt_settings_diagonal_in_tpb(expt_setting2, expt_setting4)
 
 
 def test_identity(forest):
